+++ dtslint/ts3.5/Codec.ts
@@ -106,8 +106,10 @@ _.fromSum('_tag')({
 // sum
 //
 
+const S1 = _.type({ _tag: _.literal('A'), a: _.string })
+const S2 = _.type({ _tag: _.literal('B'), b: _.number })
+
 // $ExpectType Codec<unknown, { _tag: "A"; a: string; } | { _tag: "B"; b: number; }, { _tag: "A"; a: string; } | { _tag: "B"; b: number; }>
-_.sum('_tag')({
-  A: _.type({ _tag: _.literal('A'), a: _.string }),
-  B: _.type({ _tag: _.literal('B'), b: _.number })
-})
+_.sum('_tag')({ A: S1, B: S2 })
+// // $ExpectError
+// _.sum('_tag')({ A: S1, B: S1 })