+++ dtslint/ts3.5/Eq.ts
@@ -15,3 +15,15 @@ _.partial({
     c: _.number
   })
 })
+
+//
+// sum
+//
+
+const S1 = _.type({ _tag: _.Schemable.literal('A'), a: _.string })
+const S2 = _.type({ _tag: _.Schemable.literal('B'), b: _.number })
+
+// $ExpectType Eq<{ _tag: "A"; a: string; } | { _tag: "B"; b: number; }>
+_.sum('_tag')({ A: S1, B: S2 })
+// // $ExpectError
+// _.sum('_tag')({ A: S1, B: S1 })