+++ src/Eq.ts
@@ -145,7 +145,9 @@ export const intersect = <B>(right: Eq<B>) => <A>(left: Eq<A>): Eq<A & B> => ({
  * @category combinators
  * @since 2.2.2
  */
-export function sum<T extends string>(tag: T): <A>(members: { [K in keyof A]: Eq<A[K]> }) => Eq<A[keyof A]> {
+export function sum<T extends string>(
+  tag: T
+): <A>(members: { [K in keyof A]: Eq<A[K] & Record<T, K>> }) => Eq<A[keyof A]> {
   return (members: Record<string, Eq<any>>) => {
     return {
       equals: (x: Record<string, any>, y: Record<string, any>) => {