+++ src/TaskDecoder.ts
@@ -299,7 +299,7 @@ export const record = <A>(codomain: TaskDecoder<unknown, A>): TaskDecoder<unknow
 export const fromTuple = <C extends ReadonlyArray<TaskDecoder<any, any>>>(
   ...components: C
 ): TaskDecoder<{ [K in keyof C]: InputOf<C[K]> }, { [K in keyof C]: TypeOf<C[K]> }> =>
-  K.fromTuple(M)((i, e) => FS.of(DE.index(i, DE.required, e)))(...components)
+  K.fromTuple(M)((i, e) => FS.of(DE.index(i, DE.required, e)))(...components) as any
 
 /**
  * @category combinators
@@ -351,7 +351,7 @@ export const fromSum = <T extends string>(tag: T) => <MS extends Record<string, 
  * @since 2.2.7
  */
 export const sum = <T extends string>(tag: T) => <A>(
-  members: { [K in keyof A]: TaskDecoder<unknown, A[K]> }
+  members: { [K in keyof A]: TaskDecoder<unknown, A[K] & Record<T, K>> }
 ): TaskDecoder<unknown, A[keyof A]> => pipe(UnknownRecord as any, compose(fromSum(tag)(members)))
 
 /**